+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.1] - 2025-01-28
+### Added
+- Command exception debugger
+
 ### Fixed
 - Vec2 argument type
 - Vec3 argument type